Add https instrumentation
Hook the existing http instrumentation into https as well.
