+++ packages/nodejs/src/client.ts
@@ -7,7 +7,7 @@ import { BaseMetrics } from "./metrics"
 import { NoopTracer, NoopMetrics } from "./noops"
 
 import { Instrumentation } from "./instrument"
-import * as httpPlugin from "./instrumentation/http"
+import { httpPlugin, httpsPlugin } from "./instrumentation/http"
 
 import { AppsignalOptions } from "./types/options"
 import { Plugin } from "./interfaces/plugin"
@@ -49,6 +49,7 @@ export class Client {
 
     // load plugins
     this.instrument(httpPlugin)
+    this.instrument(httpsPlugin)
   }
 
   /**
