+++ packages/nodejs/src/instrumentation/http/http.ts
@@ -1,0 +1,36 @@
+import shimmer from "shimmer"
+import http from "http"
+
+import { getPatchIncomingRequestFunction } from "./lifecycle/request"
+import { Tracer } from "../../interfaces/tracer"
+import { Plugin } from "../../interfaces/plugin"
+
+// quick alias to expose a type for the entire module
+type HttpModule = typeof http
+
+export const PLUGIN_NAME = "http"
+
+export const instrument = (
+  mod: HttpModule,
+  tracer: Tracer
+): Plugin<HttpModule> => ({
+  version: ">= 10",
+  install(): HttpModule {
+    // wrap incoming requests
+    if (mod?.Server?.prototype) {
+      shimmer.wrap(
+        mod.Server.prototype,
+        "emit",
+        getPatchIncomingRequestFunction(tracer)
+      )
+    }
+
+    return mod
+  },
+  uninstall(): void {
+    // unwrap incoming requests
+    if (mod?.Server?.prototype) {
+      shimmer.unwrap(mod.Server.prototype, "emit")
+    }
+  }
+})
